+++ src/redux/index.js
@@ -1,0 +1,27 @@
+import { combineReducers, compose, configureStore } from '@reduxjs/toolkit';
+import authReducer from '../features/auth/authSlice';
+import patientDataReducer from '../features/patient/patientDataSlice';
+import { flattenedResourcesReducer, resourceTypesReducer } from './reducers';
+import epicMiddleware, { rootEpic } from './epics';
+
+const rootReducer = combineReducers({
+  auth: authReducer,
+  patient: patientDataReducer,
+  resources: flattenedResourcesReducer,
+  resourceIdsGroupedByType: resourceTypesReducer,
+});
+
+const store = configureStore({
+  reducer: rootReducer,
+  middleware: compose([
+    epicMiddleware,
+    // routerMiddleware(history), // < e.g.: other middleware
+  ]),
+  devTools: {
+    serialize: true, // See: https://github.com/zalmoxisus/redux-devtools-extension/blob/master/docs/API/Arguments.md#serialize
+  },
+});
+
+epicMiddleware.run(rootEpic);
+
+export default store;
